requirements.txt
requests>=2.32.4,<3.0.0
pandas>=2.2.3
google-genai>=1.21.1,<2.0.0
google-adk==1.14.1
python-dotenv>=1.0.0,<2.0.0
colorama
